stop trying to kill dead things and reanimate alive things

+++ theme/themes/reload.sh
@@ -1,12 +1,12 @@
-#!/bin/env bash
+#!/usr/bin/env bash
 
-killall lemonbar
-killall compton
-killall -USR1 termite
+[ ! -z $(pgrep lemonbar) ] && killall lemonbar
+[ ! -z $(pgrep picom) ] && killall picom
+# killall -USR1 alacritty
 
 if [ $(pgrep qutebrowser | wc -l) -gt 0 ] ; then
     qutebrowser ':config-source'
 fi
-bash $HOME/.config/bspwm/bspwmrc
+$HOME/.config/bspwm/bspwmrc
 
 spicetify update
